eCommerce Application/starter_code/src/main/java/com/example/demo/controllers/UserController.java
package com.example.demo.controllers;

import java.security.MessageDigest;
import java.security.NoSuchAlgorithmException;
import java.util.List;
import java.util.Optional;

import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import com.example.demo.model.persistence.Cart;
import com.example.demo.model.persistence.User;
import com.example.demo.model.persistence.repositories.CartRepository;
import com.example.demo.model.persistence.repositories.UserRepository;
import com.example.demo.model.requests.CreateUserRequest;


@Slf4j
@RestController
@RequestMapping("/api/user")
public class UserController {
	
	@Autowired
	private UserRepository userRepository;
	
	@Autowired
	private CartRepository cartRepository;

	@Autowired
	private BCryptPasswordEncoder bCryptPasswordEncoder;



	@GetMapping("/all")
	public ResponseEntity<  List<User> > findAll() {
		return ResponseEntity.of(  Optional.of(userRepository.findAll()) );

	}


	@GetMapping("/id/{id}")
	public ResponseEntity<User> findById(@PathVariable Long id) {
		return ResponseEntity.of(userRepository.findById(id));
	}



	@GetMapping("/{username}")
	public ResponseEntity<User> findByUserName(@PathVariable String username) {
		User user = userRepository.findByUsername(username);
		return user == null ? ResponseEntity.notFound().build() : ResponseEntity.ok(user);
	}




	@PostMapping("/create")
	public ResponseEntity<User> createUser(@RequestBody CreateUserRequest createUserRequest) {

		log.info("call on createUser");

		User user = new User();
		user.setUsername(createUserRequest.getUsername());
		Cart cart = new Cart();
		cartRepository.save(cart);
		user.setCart(cart);
		if(createUserRequest.getPassword().length()<7 ||
				!createUserRequest.getPassword().equals(createUserRequest.getConfirmPassword())){
			//System.out.println("Error - Either length is less than 7 or pass and conf pass do not match. Unable to create ",
			//		createUserRequest.getUsername());
			return ResponseEntity.badRequest().build();
		}

		//Hashing the password !!
		String passwordToHash = createUserRequest.getPassword();

		//System.out.println("passwordToHash="+passwordToHash);
		String securePassword = get_SecurePassword(passwordToHash, user.getSalt(), "MD5");
		//System.out.println("securePassword="+securePassword);

		String encodedPassword =bCryptPasswordEncoder.encode(securePassword);
		user.setPassword(encodedPassword);
		//System.out.println("encodedPassword="+encodedPassword);

		userRepository.save(user);
		return ResponseEntity.ok(user);
	}







	// Method to generate the hash.
	//It takes a password and the Salt as input arguments
	private static String get_SecurePassword(String passwordToHash, byte[] salt, String algorithm ){
		String generatedPassword = null;
		try {
			MessageDigest md = MessageDigest.getInstance(algorithm);

			if(salt!=null) {
				md.update(salt);
			}

			byte[] bytes = md.digest(passwordToHash.getBytes());
			StringBuilder sb = new StringBuilder();
			for(int i=0; i< bytes.length ;i++)
			{
				sb.append(Integer.toString((bytes[i] & 0xff) + 0x100, 16).substring(1));
			}
			generatedPassword = sb.toString();
		}
		catch (NoSuchAlgorithmException e) {
			e.printStackTrace();
		}
		return generatedPassword;
	}







}
